Enable compiler warnings (and fix existing issues) (#24)
* Enable some compiler warnings


* Fix narrowing conversion warnings


* Fix struct initialization warnings

+++ everestjs/everestjs.cpp
@@ -395,7 +395,7 @@ static Napi::Value boot_module(const Napi::CallbackInfo& info) {
             }
             auto req_mod_vars_array = Napi::Array::New(env);
             auto req_mod_cmds_array = Napi::Array::New(env);
-            for (int i = 0; i < req_route_list.size(); i++) {
+            for (size_t i = 0; i < req_route_list.size(); i++) {
                 auto req_route = req_route_list[i];
                 const std::string& requirement_module_id = req_route["module_id"];
                 const std::string& requirement_impl_id = req_route["implementation_id"];
